pyproject.toml
[project]
name = "cold-mailer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.13.0.2",
    "google>=3.0.0",
    "google-api-python-client>=2.186.0",
    "google-auth>=2.42.1",
    "google-auth-httplib2>=0.2.1",
    "google-auth-oauthlib>=1.2.2",
    "python-docx>=1.2.0",
]
